+++ CHANGELOG.md
@@ -1,6 +1,11 @@
-## 1.0.0/2025-07-07
-- initial release
+## 2.1.0/2025-10-01
+- duration support
+- legacy ascii encoding -> latin1
+- property allValue -> extra in decode()
 ## 2.0.0/2025-07-09
 - changed command syntax to increase compatibility 
   ('?' directly after parameter name, comma as separator for values)
+## 1.0.0/2025-07-07
+- initial release
 ## TODO:
+- parameter dictionary. Comma separated aliases -> [string]